requirements.txt
torch>=2.0.0
numpy>=1.21.0
transformers>=4.20.0
wandb>=0.13.0
accelerate>=0.20.0
datasets>=2.0.0
tokenizers>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
omegaconf>=2.3.0
einops>=0.6.0
psutil>=5.9.0

# Optional: Flash Attention for performance (requires CUDA)
# flash-attn>=2.0.0

# Optional: xFormers for efficient attention (alternative to Flash Attention)
# xformers>=0.0.20

# Optional: Accelerate for distributed training (recommended)
# accelerate>=0.20.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.0.0
pytest-mock>=3.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991